fix(analytics): address PR #479 review — helper-module split, tooltip scaling, header
- zeroFillDays/DayPointValue move to lib/daySeries.ts (JSX-free): the
  screen's static import no longer creates a synchronous edge into
  AnalyticsCharts, so next/dynamic actually keeps the chart code out of
  the eager chunk. AnalyticsCharts re-exports for a single import site.
- Tooltip positions as a wrapper PERCENTAGE (tooltipLeftPct, tested)
  instead of viewBox pixels — no drift when the rendered width falls
  under the 240px measurement floor (narrow viewports).
- Stale #121-era screen header rewritten; unused type import dropped.
